Moves all functionality of txes mocking to 'TxsMocker' + updates tests
